BACKLOG-16051 Remove unused file, fix renaming, update implementation
