+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentLayout.utils.js
@@ -52,14 +52,14 @@ export const structureData = function (parentPath, dataForParentPath = []) {
     }
 };
 
-export const flattenTree = function (rows, selector = p => p) {
+export const flattenTree = function (rows) {
     const items = [];
     collectItems(rows);
     return items;
 
     function collectItems(arrayData) {
         for (let i = 0; i < arrayData.length; i++) {
-            items.push(selector(arrayData[i]));
+            items.push(arrayData[i]);
             collectItems(arrayData[i].subRows || []);
         }
     }
